requirements.txt
absl-py==2.2.2
altair==4.2.2
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
flatbuffers==1.12
fonttools==4.57.0
gast==0.4.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==3.19.6
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
seaborn==0.13.2
semver==3.0.4
six==1.17.0
smmap==5.0.2
soupsieve==2.6
stack-data==0.6.3
streamlit==1.11.0
tenacity==9.1.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
